funny0money/requirements.txt
line-bot-sdk==1.8.0
flask==1.0.2
gunicorn
